refactor(self-custodial): move the account-mode surfaces into the self-custodial namespace
