--- app/components/anon-mode-convert-modal/index.ts
@@ -1,1 +1,0 @@
-export { AnonModeConvertModal } from "./anon-mode-convert-modal"
