+++ app/screens/settings-screen/self-custodial/account-mode.tsx
