+++ app/screens/settings-screen/settings-screen.tsx
@@ -29,7 +29,7 @@ import { PhoneSetting } from "./account/settings/phone"
 import { SettingsGroup } from "./group"
 import { DefaultWallet } from "./settings/account-default-wallet"
 import { AccountLevelSetting } from "./settings/account-level"
-import { AccountModeSetting } from "./settings/account-mode"
+import { AccountModeSetting } from "./self-custodial/account-mode"
 import { AccountLNAddress } from "./settings/account-ln-address"
 import { PhoneLnAddress } from "./settings/phone-ln-address"
 import { AccountPOS } from "./settings/account-pos"
